requirements.txt
python-dotenv
hupper
requests
openai
langchain
langchain-experimental
gradio
flask
transformers
pandas
neo4j
atlassian-python-api
mysql-connector-python